requirements.txt
absl-py==0.2.2
alabaster==0.7.9
anaconda-clean==1.0
anaconda-client==1.5.1
anaconda-navigator==1.3.1
argcomplete==1.0.0
args==0.1.0
astor==0.6.2
astroid==1.4.7
astropy==1.2.1
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
beautifulsoup4==4.5.1
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.2
boto==2.42.0
Bottleneck==1.1.0
cachetools==2.1.0
certifi==2018.1.18
cffi==1.7.0
chardet==3.0.4
ChatterBot==0.8.4
chatterbot-corpus==1.1.1
chest==0.2.3
click==6.7
clint==0.5.1
cloudpickle==0.2.1
clyent==1.2.2
colorama==0.3.7
conda==4.3.29
conda-build==2.0.2
config==0.3.9
configobj==5.0.6
contextlib2==0.5.3
cookies==2.2.1
cryptography==1.5
cycler==0.10.0
Cython==0.24.1
cytoolz==0.8.0
dask==0.11.0
datashape==0.5.2
decorator==4.0.10
dill==0.2.5
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.10
django-bootstrap3==9.1.0
django-ckeditor==5.4.0
django-configurations==2.0
django-crispy-forms==1.7.0
django-ddp==0.19.1
django-debug-panel==0.8.3
django-debug-toolbar==1.8
django-extensions==1.9.7
django-formtools==2.0
django-js-asset==0.1.1
django-multiselectfield==0.1.8
django-pagedown==0.1.3
django-recaptcha==1.3.1
django-registration-redux==2.2
django-reversion==2.0.10
django-skd-smoke==0.2
django-slack-invitation==0.1.0
django-smart-selects==1.5.2
django-social-widgets==0.4.0
django-sslserver==0.20
django-toolbelt==0.0.1
docutils==0.12
dynd==0.7.3.dev1
enum34==1.1.6
et-xmlfile==1.0.1
everett==0.9
fastcache==1.0.2
filelock==2.0.6
flake8==3.5.0
Flask==0.12.2
Flask-Cors==3.0.2
floyd-cli==0.10.36
future==0.16.0
fuzzywuzzy==0.15.1
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gast==0.2.0
gcloud==0.18.3
geopy==1.11.0
gevent==1.2.1
gevent-websocket==0.10.1
google==2.0.1
google-api-core==0.1.4
google-auth==1.4.2
google-cloud==0.33.1
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.1
google-cloud-core==0.28.1
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.2
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.1
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.0.1
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
greenlet==0.4.10
grpc-google-iam-v1==0.11.4
grpcio==1.12.0
gunicorn==19.1.1
h5py==2.7.1
HeapDict==1.0.0
html5lib==0.9999999
httplib2==0.9.2
idna==2.6
image==1.3.3
imagesize==0.7.1
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jsonfield==2.0.2
jsonpickle==0.9.6
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
kaggle==1.1.0
Keras==2.0.2
lazy-object-proxy==1.2.1
llvmlite==0.13.0
locket==0.2.0
lxml==3.6.4
Markdown==2.6.11
MarkupSafe==0.23
marshmallow==2.15.0
martor==1.2.5
mathparse==0.1.1
matplotlib==1.5.3
mccabe==0.6.1
medium==0.3.0
meteor-ejson==1.1.0
mistune==0.7.3
mpmath==0.19
multipledispatch==0.4.8
nb-anacondacloud==1.2.0
nb-conda==2.0.0
nb-conda-kernels==2.0.0
nbconvert==4.2.0
nbformat==4.3.0
nbpresent==3.0.2
networkx==1.11
nltk==3.2.1
nose==1.3.7
notebook==4.2.3
numba==0.28.1
numexpr==2.6.1
numpy==1.14.3
oauth2client==3.0.0
oauthlib==2.0.6
odo==0.5.0
olefile==0.44
opencv-contrib-python==3.3.0.10
opencv-python==3.3.0.10
openpyxl==2.3.2
pandas==0.18.1
partd==0.3.6
path.py==0.0.0
pathlib2==2.3.0
patsy==0.4.1
pbr==3.1.1
pep8==1.7.0
pew==1.1.0
pexpect==4.0.1
pickleshare==0.7.4
Pillow==4.3.0
pipenv==8.2.7
pkginfo==1.3.2
plotly==2.0.13
pluggy==0.6.0
ply==3.8
prompt-toolkit==1.0.3
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.2.post1
psutil==5.4.5
psycogreen==1.0
psycopg2==2.7.3.2
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.2.1
pybars3==0.9.3
pycodestyle==2.3.1
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pydot==1.2.3
pyflakes==1.6.0
Pygments==2.1.3
PyJWT==1.5.2
pylint==1.5.4
PyMeta3==0.5.1
pymongo==3.6.1
pyOpenSSL==16.2.0
pyparsing==2.1.4
pytest==2.9.2
python-dateutil==2.6.1
python-dotenv==0.7.1
python-forecastio==1.3.4
python-twitter==3.4.1
pytz==2018.3
PyYAML==3.12
pyzmq==15.4.0
QtAwesome==0.3.3
qtconsole==4.2.1
QtPy==1.1.2
RapidConnect==0.0.6
raven==6.6.0
redis==2.10.5
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
responses==0.8.1
reverse-engineer==1.1
reverse-geocode==1.0
rope-py3k==0.9.4.post1
rsa==3.4.2
ruamel-yaml===-VERSION
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==1.1.0
selenium==3.8.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
slackipycore==0.0.4
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.4.6
spyder==3.0.0
SQLAlchemy==1.2.5
sqlparse==0.2.3
static3==0.5.1
statsmodels==0.6.1
stevedore==1.28.0
sympy==1.0
tables==3.2.3.1
tabulate==0.8.2
tensorboard==1.8.0
tensorflow==1.4.0
tensorflow-hub==0.1.0
tensorflow-tensorboard==0.4.0
tensorflowjs==0.4.0
termcolor==1.1.0
terminado==0.6
Theano==1.0.0
toolz==0.8.0
tornado==4.4.1
tox==2.9.1
traitlets==4.3.0
tweepy==3.5.0
twython==3.6.0
typing==3.6.2
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.22
uWSGI==2.0.15
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
wcwidth==0.1.7
websocket-client==0.47.0
Werkzeug==0.11.11
whitenoise==3.3.1
widgetsnbextension==1.2.6
wit==5.1.0
wrapt==1.10.6
xadmin==0.6.1
xlrd==1.0.0
XlsxWriter==0.9.3
xlwt==1.2.0
